refactor(entitlements): move SubscriptionValues to its entity file
