+++ src/modules/entitlements/domain/subscriptions.repository.interface.ts
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: FSL-1.1-MIT
 import type { EntityManager } from 'typeorm';
-import type { SpaceSubscription } from '@/modules/entitlements/datasources/entities/space-subscription.entity.db';
+import type { SubscriptionValues } from '@/modules/entitlements/domain/entities/space-subscription.entity';
 import type { Space } from '@/modules/spaces/domain/entities/space.entity';
 
 export const ISubscriptionsRepository = Symbol('ISubscriptionsRepository');
@@ -53,13 +53,3 @@ export interface ISubscriptionsRepository {
     entityManager: EntityManager,
   ): Promise<void>;
 }
-
-export type SubscriptionValues = Pick<
-  SpaceSubscription,
-  | 'status'
-  | 'planId'
-  | 'planName'
-  | 'currentPeriodStart'
-  | 'currentPeriodEnd'
-  | 'lastEventAt'
->;
